moved convertToArray and updated type
